% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{calc_cw_mean}
\alias{calc_cw_mean}
\alias{calc_cw_variance}
\alias{calc_cw_skewness}
\alias{calc_cw_kurtosis}
\alias{calc_cw_moments}
\title{Calculate moments of trait distribution}
\usage{
calc_cw_mean(trait = NULL, weight = NULL)

calc_cw_variance(trait = NULL, weight = NULL, cw_mean = NULL)

calc_cw_skewness(
  trait = NULL,
  weight = NULL,
  cw_mean = NULL,
  cw_variance = NULL
)

calc_cw_kurtosis(
  trait = NULL,
  weight = NULL,
  cw_mean = NULL,
  cw_variance = NULL
)

calc_cw_moments(trait = NULL, weight = NULL)
}
\arguments{
\item{trait}{numeric vector of trait values}

\item{weight}{numeric vector of relative abundance or biomass}
}
\value{
a numeric vector
}
\description{
Calculate moments of trait distribution
}
\examples{
#vector of species abundance
abun <- rpois(n = 10, lambda = 3)
# vector of species trait
trait <- rnorm(n = 10, mean = 10, sd = 3)

calc_cw_moments(trait = trait, weight = abun)
}
