% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_plant_interactions.R
\name{check_int_arg}
\alias{check_int_arg}
\title{Check interaction index data}
\usage{
check_int_arg(without, with_nbs)
}
\arguments{
\item{without}{data from the without}

\item{with_nbs}{data from the without}
}
\description{
Check if the arguments are in the good format to compute the index. Returns
NA if with_nbs or/and without is NA
}
