% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_stability.R
\name{community_stability}
\alias{community_stability}
\alias{statistical_averaging_effect}
\alias{compensatory_effect}
\alias{asynchrony}
\alias{population_stability}
\title{Community stability decomposition}
\usage{
community_stability(mat)

statistical_averaging_effect(mat)

compensatory_effect(mat)

asynchrony(mat)

population_stability(mat)
}
\arguments{
\item{mat}{a time x species matrix containing abundances, or biomasses,
etc...}
}
\description{
Community stability decomposition
}
\details{
Add the paper references
}
