% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{scale_weight}
\alias{scale_weight}
\title{Scale weight}
\usage{
scale_weight(weight = NULL)
}
\arguments{
\item{weight}{numeric vector of relative abundance or biomass}
}
\description{
Divide weights by their sum
}
\examples{
abun <- rpois(n = 10, lambda = 3)
scaled <- scale_weight(abun)
c(base = sum(abun), scaled = scaled)
}
